refractor(cli): Remade the entire cli to a better version and simplier version
